Order/take

+++ src/modules/order/order.service.ts
@@ -319,7 +319,14 @@ export class OrderService {
       new: true,
     });
   }
-  removeDiscount(id: number) {
+  async removeDiscount(id: number) {
+    const orders = await this.orderModel.find({ discount: id });
+    if (orders.length > 0) {
+      throw new HttpException(
+        'Discount is used in orders',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
     return this.discountModel.findByIdAndRemove(id);
   }
   async createOrderForDivide(
@@ -394,7 +401,8 @@ export class OrderService {
       orderId: number;
     }[],
     discount: number,
-    discountPercentage: number,
+    discountPercentage?: number,
+    discountAmount?: number,
   ) {
     for (const orderItem of orders) {
       const oldOrder = await this.orderModel.findById(orderItem.orderId);
@@ -405,7 +413,15 @@ export class OrderService {
         try {
           await this.orderModel.findByIdAndUpdate(orderItem.orderId, {
             discount: discount,
-            discountPercentage: discountPercentage,
+            ...(discountPercentage && {
+              discountPercentage: discountPercentage,
+            }),
+            ...(discountAmount && {
+              discountAmount: Math.min(
+                discountAmount,
+                oldOrder.unitPrice * orderItem.selectedQuantity,
+              ),
+            }),
           });
         } catch (error) {
           throw new HttpException(
@@ -421,7 +437,15 @@ export class OrderService {
           ...orderDataWithoutId,
           quantity: orderItem.selectedQuantity,
           discount: discount,
-          discountPercentage: discountPercentage,
+          ...(discountPercentage && {
+            discountPercentage: discountPercentage,
+          }),
+          ...(discountAmount && {
+            discountAmount: Math.min(
+              discountAmount,
+              orderDataWithoutId.unitPrice * orderItem.selectedQuantity,
+            ),
+          }),
           paidQuantity: 0,
         });
         try {
